Hopefully the last of major edits.
This set of edits simplified the structure of the leaf nodes by moving
some of the common internal variables to ResourceImpl, the bottom of the
hierarchy.

I also eliminated some duplication in methods such as
getMemoryRequestServer, as well as simplified some of the check methods.
