+++ datasketches-memory-java8/src/main/java/org/apache/datasketches/memory/internal/BBWritableMemoryImpl.java
@@ -35,10 +35,6 @@ final class BBWritableMemoryImpl extends NativeWritableMemoryImpl {
   private final ByteBuffer byteBuf; //holds a reference to a ByteBuffer until we are done with it.
   private final Object unsafeObj;
   private final long nativeBaseOffset; //raw off-heap address of allocation base if ByteBuffer direct, else 0
-  private final long offsetBytes;      //from the root resource including original ByteBuffer position or region offset
-  private final long capacityBytes;
-  private final int typeId;
-  private long cumOffsetBytes;         //includes array header if heap, and nativeBaseOffset if off-heap
 
   BBWritableMemoryImpl(
       final Object unsafeObj,
@@ -56,7 +52,7 @@ final class BBWritableMemoryImpl extends NativeWritableMemoryImpl {
     this.capacityBytes = capacityBytes;
     this.typeId = removeNnBuf(typeId) | BYTEBUF | MEMORY | NATIVE;
     this.cumOffsetBytes = cumOffsetBytes;
-    this.memReqSvr = memReqSvr;
+    this.memReqSvr = memReqSvr; //in ResourceImpl
     this.byteBuf = byteBuf;
     if ((this.owner != null) && (this.owner != Thread.currentThread())) {
       throw new IllegalStateException(THREAD_EXCEPTION_TEXT);
@@ -103,31 +99,6 @@ final class BBWritableMemoryImpl extends NativeWritableMemoryImpl {
   @Override
   public ByteBuffer getByteBuffer() {
     return byteBuf;
-  }
-
-  @Override
-  public long getCapacity() {
-    return capacityBytes;
-  }
-
-  @Override
-  public long getCumulativeOffset() {
-    return cumOffsetBytes;
-  }
-
-  @Override
-  public long getNativeBaseOffset() {
-    return nativeBaseOffset;
-  }
-
-  @Override
-  public long getTotalOffset() {
-    return offsetBytes;
-  }
-
-  @Override
-  int getTypeId() {
-    return typeId;
   }
 
   @Override
